Move impl Contains for Netv6Addr into netv6addr mod
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>
